Move to targeting based on pooled env name
- use smith rather than fly and claim job output
- use env name rather than job name
- use non-interactive login

+++ toolsmiths/claim-pooled-env/target-claimed-env.sh
@@ -2,20 +2,21 @@
 # source this file to target your interactive shell
 
 # Optional env overrides:
-# CONCOURSE_TARGET if other than 'denver'
 # OM_PATH to save somewhere other than /tmp
 
 source $(dirname $BASH_SOURCE)/../../helpers/toolsmiths.sh
 
 if [ $# -ne 1 ]; then
-    echo "Usage: $0 <concourse-job>"
-    return
+    echo "Usage: $0 <Pooled Env Name>"
+    return 1
 fi
 
-fly -t ${CONCOURSE_TARGET:-denver} status 2>/dev/null || fly -t ${CONCOURSE_TARGET:-denver} login -b
-fly -t ${CONCOURSE_TARGET:-denver} w -j ${1} | tail -1  | sed 's/succeeded$//' >/tmp/claimed-metadata
-export ENV_NAME=$(jq .name -r /tmp/claimed-metadata)
+export ENV_NAME=$1
 echo "Targeting ${ENV_NAME}."
+
+if ! smith read -e ${ENV_NAME} > /tmp/claimed-metadata; then
+  return 1;
+fi
 
 export ENV_METADATA="/tmp/${ENV_NAME}-metadata"
 mv /tmp/claimed-metadata ${ENV_METADATA}
@@ -24,7 +25,9 @@ bosh-login ${ENV_METADATA}
 
 export CF_ADMIN_PASSWORD=$(om credentials -p cf -c .uaa.admin_credentials -f password)
 export CF_ADMIN_USER=$(om credentials -p cf -c .uaa.admin_credentials -f identity)
-cf login -a https://api.$(jq .sys_domain -r ${ENV_METADATA}) -u ${CF_ADMIN_USER} -p ${CF_ADMIN_PASSWORD} --skip-ssl-validation
+
+cf api https://api.$(jq .sys_domain -r ${ENV_METADATA}) --skip-ssl-validation
+cf auth ${CF_ADMIN_USER} ${CF_ADMIN_PASSWORD}
 
 jq .ops_manager.password -r ${ENV_METADATA} | pbcopy
 echo "OpsMan password copied, log in at: $(jq .ops_manager.url -r ${ENV_METADATA})"
